Try not to wipe out git history
