--- packages/hash/types/Hash.d.ts
@@ -1,3 +1,0 @@
-import { Options, HashHistory } from "./types";
-declare function Hash<Q>(options?: Options<Q>): HashHistory<Q>;
-export { Hash };
