+++ packages/hash/types/index.d.ts
@@ -1,3 +1,3 @@
-import { Hash } from "./Hash";
+import { Options, HashHistory } from "./types";
 export * from "./types";
-export { Hash };
+export declare function Hash<Q>(options?: Options<Q>): HashHistory<Q>;
